trim_bag の実装をもとに戻した
